refactored DAO-s: resource dao do not call each-other methods any more. Only the new Dispatcher method can do it. Not tested.
git-svn-id: https://svn.clarin.eu/DASISH/t5.6/backend/annotator-backend/trunk/annotator-backend@3407 72633d89-856b-0410-a642-59178932f03a

+++ src/main/java/eu/dasish/annotation/backend/dao/impl/JdbcAnnotationDao.java
@@ -19,21 +19,12 @@ package eu.dasish.annotation.backend.dao.impl;
 
 import eu.dasish.annotation.backend.Helpers;
 import eu.dasish.annotation.backend.dao.AnnotationDao;
-import eu.dasish.annotation.backend.dao.NotebookDao;
-import eu.dasish.annotation.backend.dao.PermissionsDao;
-import eu.dasish.annotation.backend.dao.SourceDao;
-import eu.dasish.annotation.backend.dao.UserDao;
 import eu.dasish.annotation.backend.identifiers.AnnotationIdentifier;
-import eu.dasish.annotation.backend.identifiers.SourceIdentifier;
-import eu.dasish.annotation.backend.identifiers.UserIdentifier;
 import eu.dasish.annotation.schema.Annotation;
-import eu.dasish.annotation.schema.AnnotationBody;
 import eu.dasish.annotation.schema.AnnotationInfo;
 import eu.dasish.annotation.schema.NewOrExistingSourceInfo;
 import eu.dasish.annotation.schema.NewOrExistingSourceInfos;
-import eu.dasish.annotation.schema.NewSourceInfo;
 import eu.dasish.annotation.schema.ResourceREF;
-import eu.dasish.annotation.schema.Source;
 import eu.dasish.annotation.schema.SourceInfo;
 import java.lang.String;
 import java.sql.ResultSet;
@@ -45,7 +36,6 @@ import java.util.List;
 import java.util.Map;
 import javax.sql.DataSource;
 import javax.xml.datatype.DatatypeConfigurationException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.RowMapper;
 
 /**
@@ -54,15 +44,6 @@ import org.springframework.jdbc.core.RowMapper;
  * @author Peter Withers <peter.withers@mpi.nl>
  */
 public class JdbcAnnotationDao extends JdbcResourceDao implements AnnotationDao {
-
-    @Autowired
-    PermissionsDao jdbcPermissionsDao;
-    @Autowired
-    UserDao jdbcUserDao;
-    @Autowired
-    NotebookDao jdbcNotebookDao;
-    @Autowired
-    SourceDao jdbcSourceDao;
 
     public JdbcAnnotationDao(DataSource dataSource) {
         setDataSource(dataSource);
@@ -86,26 +67,25 @@ public class JdbcAnnotationDao extends JdbcResourceDao implements AnnotationDao 
 
     ////////////////////////////////////////////////////////////////////////
     @Override
-    public List<Number> getFilteredAnnotationIDs(String link, String text, String access, String namespace, UserIdentifier owner, Timestamp after, Timestamp before) {
+    public List<Number> getFilteredAnnotationIDs(List<Number> annotationIDs, String text, String access, String namespace, Number ownerID, Timestamp after, Timestamp before) {
 
         StringBuilder sql = new StringBuilder("SELECT DISTINCT ");
         sql.append(annotation_id).append(" FROM ").append(annotationTableName).append(" WHERE TRUE ");
         Map<String, Object> params = new HashMap<String, Object>();
 
-        if (link != null) {
-            List<Number> sourceIDs = jdbcSourceDao.getSourcesForLink(link);
-            List<Number> annotationIDs = retrieveAnnotationList(sourceIDs);
-            if (!annotationIDs.isEmpty()) {
-                String values = makeListOfValues(annotationIDs);
-                sql.append(" AND ").append(annotation_id).append(" IN ").append(values);
-            } else {
+        if (annotationIDs == null) {
+            return null;
+        } else {
+            if (annotationIDs.isEmpty()) {
                 return new ArrayList<Number>();
             }
         }
 
+        String values = makeListOfValues(annotationIDs);
+        sql.append(" AND ").append(annotation_id).append(" IN ").append(values);
 
-        if (owner != null) {
-            Number ownerID = jdbcUserDao.getInternalID(owner);
+
+        if (ownerID != null) {
             sql.append(" AND ").append(owner_id).append(" = :owner ");
             params.put("owner", ownerID);
         }
@@ -123,7 +103,6 @@ public class JdbcAnnotationDao extends JdbcResourceDao implements AnnotationDao 
         if (text != null) {
             sql.append(" AND ").append(body_xml).append("  LIKE '%").append(text).append("%'");
         }
-
 
         List<Number> result = getSimpleJdbcTemplate().query(sql.toString(), internalIDRowMapper, params);
         return result;
@@ -205,76 +184,74 @@ public class JdbcAnnotationDao extends JdbcResourceDao implements AnnotationDao 
     };
 
     //////////////////////////////////////////////////////////////////////////
-    @Override
-    public Annotation getAnnotation(Number annotationID) throws SQLException {
+    @Override    
+    public Annotation getAnnotationWithoutSources(Number annotationID) throws SQLException {
         if (annotationID == null) {
-            return null;
-        }
+            return null;        }
         String sql = "SELECT " + annotationStar + " FROM " + annotationTableName + " WHERE " + annotationAnnotation_id + "= ? LIMIT  1";
-        List<Annotation> result = getSimpleJdbcTemplate().query(sql, annotationRowMapper, annotationID);
+        List<Annotation> respond = getSimpleJdbcTemplate().query(sql, annotationRowMapper, annotationID);
 
-        if (result == null) {
+        if (respond == null) {
             return null;
         }
-        if (result.isEmpty()) {
-            return null;
-        }
-        return result.get(0);
+      
+        return respond.get(0);
     }
+    
     private final RowMapper<Annotation> annotationRowMapper = new RowMapper<Annotation>() {
         @Override
         public Annotation mapRow(ResultSet rs, int rowNumber) throws SQLException {
-            Annotation result = new Annotation();
+            
+            Map<String, Object> result = new HashMap<String, Object>();
+            
+            Annotation annotation = new Annotation();
 
             ResourceREF ownerREF = new ResourceREF();
             ownerREF.setRef(String.valueOf(rs.getInt(owner_id)));
-            result.setOwner(ownerREF);
+            annotation.setOwner(ownerREF);
 
-            result.setHeadline(rs.getString(headline));
+            annotation.setHeadline(rs.getString(headline));
 
-            result.setBody(Helpers.deserializeBody(rs.getString(body_xml)));
-
-            List<SourceInfo> sourceInfoList = jdbcSourceDao.getSourceInfos(retrieveSourceIDs(rs.getInt(annotation_id)));
-            NewOrExistingSourceInfos noeSourceInfos = contructNewOrExistingSourceInfo(sourceInfoList);
-            result.setTargetSources(noeSourceInfos);
+            annotation.setBody(Helpers.deserializeBody(rs.getString(body_xml)));
+            annotation.setTargetSources(null);
 
             // TODO: fix: rpelace URI in the schema with external id, or make here the conversion:
             // from external ID in the DB to the URI for the class
-            result.setURI(rs.getString(external_id));
+            annotation.setURI(rs.getString(external_id));
 
             try {
-                result.setTimeStamp(Helpers.setXMLGregorianCalendar(rs.getTimestamp(time_stamp)));
-                return result;
+                annotation.setTimeStamp(Helpers.setXMLGregorianCalendar(rs.getTimestamp(time_stamp)));
+                return annotation;
             } catch (DatatypeConfigurationException e) {
                 System.out.println(e);
-                return result; // no date-time is set 
+                return annotation; // no date-time is set 
             }
         }
     };
-
+    
+    
+   
+    //////////////////////////////////////////////////////
     @Override
-    public int[] deleteAnnotation(Number annotationId) throws SQLException {
+    public int[] deleteAnnotation(Number annotationID) throws SQLException {
 
-        int[] result = new int[5];
-
-        String sqlNotebooks = "DELETE FROM " + notebooksAnnotationsTableName + " where " + annotation_id + " = ?";
-        result[0] = getSimpleJdbcTemplate().update(sqlNotebooks, annotationId);// removed "notebooks_annotations" rows 
-
-        String sqlPermissions = "DELETE FROM " + permissionsTableName + " where " + annotation_id + " = ?";
-        result[1] = getSimpleJdbcTemplate().update(sqlPermissions, annotationId); // removed "permission" rows 
-
-        // safe removing sources
-        List<Number> sourceIDs = retrieveSourceIDs(annotationId);
-        String sqlTargetSources = "DELETE FROM " + annotationsSourcesTableName + " where " + annotation_id + " = ?";
-        result[2] = getSimpleJdbcTemplate().update(sqlTargetSources, annotationId); // removed "annotations_target_sources" rows
-        result[3] = 0; //removed "target_source" rows 
-        for (Number sourceID : sourceIDs) {
-            int[] deleteSource = jdbcSourceDao.deleteSource(sourceID);
-            result[3] = result[3] + deleteSource[1];
+        int[] result = new int[3];
+        
+        if (annotationIsInUse(annotationID)) {
+            result[0] =0;
+            result[1] =0;
+            result[2] =0;
+            return result;
         }
-
+        
+        String sqlPermissions = "DELETE FROM " + permissionsTableName + " where " + annotation_id + " = ?";
+        result[0] = getSimpleJdbcTemplate().update(sqlPermissions, annotationID); // removed "permission" rows 
+        
+        String sqlTargetSources = "DELETE FROM " + annotationsSourcesTableName + " where " + annotation_id + " = ?";
+        result[1] = getSimpleJdbcTemplate().update(sqlTargetSources, annotationID); // removed "annotations_target_sources" rows
+       
         String sqlAnnotation = "DELETE FROM " + annotationTableName + " where " + annotation_id + " = ?";
-        result[4] = getSimpleJdbcTemplate().update(sqlAnnotation, annotationId); // removed annotations rows
+        result[2] = getSimpleJdbcTemplate().update(sqlAnnotation, annotationID); // removed annotations rows
 
         return result;
     }
@@ -283,7 +260,6 @@ public class JdbcAnnotationDao extends JdbcResourceDao implements AnnotationDao 
     // Change it when the decision is taken!!!
     @Override
     public Number addAnnotation(Annotation annotation, Number ownerID) throws SQLException {
-
         // generate a new annotation ID 
         AnnotationIdentifier annotationIdentifier = new AnnotationIdentifier();
         Map<String, Object> params = new HashMap<String, Object>();
@@ -325,56 +301,36 @@ public class JdbcAnnotationDao extends JdbcResourceDao implements AnnotationDao 
         return result;
     }
 
-    /**
-     *
-     * @param annotationID
-     * @return mapping of the column names of the DB to the corresponding values
-     * used for testing
-     */
-    //NOT TESTED
-    public Map<String, Object> getRawAnnotation(Number annotationID) throws SQLException {
-
-        if (annotationID == null) {
-            return null;
-        }
-        String sql = "SELECT " + annotationStar + " FROM " + annotationTableName + " WHERE " + annotationAnnotation_id + "= ? LIMIT  1";
-        List<Map<String, Object>> result = getSimpleJdbcTemplate().query(sql, rawAnnotationRowMapper, annotationID);
-
-        if (result == null) {
-            return null;
-        }
-        if (result.isEmpty()) {
-            return null;
-        }
-
-        return result.get(0);
+  
+    //////////////////////////////////////////////////////////////////////////////////
+    @Override
+    public int addAnnotationSourcePair(Number annotationID, Number sourceID) throws SQLException {
+        Map<String, Object> paramsAnnotationsSources = new HashMap<String, Object>();
+        paramsAnnotationsSources.put("annotationId", annotationID);
+        paramsAnnotationsSources.put("sourceId", sourceID);
+        String sqlAnnotationsSources = "INSERT INTO " + annotationsSourcesTableName + "(" + annotation_id + "," + source_id + " ) VALUES (:annotationId, :sourceId)";
+        int affectedRows = getSimpleJdbcTemplate().update(sqlAnnotationsSources, paramsAnnotationsSources);
+        return (affectedRows);
     }
-    private final RowMapper<Map<String, Object>> rawAnnotationRowMapper = new RowMapper<Map<String, Object>>() {
+    //////////////////////////////////////////////////////////////////////////////////
+
+    /////////////// helpers //////////////////
+    
+        //////////////////////////////
+    private boolean annotationIsInUse(Number sourceID) {
+        String sql = "SELECT " + notebook_id + " FROM " + notebooksAnnotationsTableName + " WHERE " + annotation_id + "= ? LIMIT 1";
+        List<Number> result = getSimpleJdbcTemplate().query(sql, notebookIDRowMapper, sourceID);
+        if (result.size() > 0) {
+            return true;
+        }
+        return false;
+    }
+    
+      private final RowMapper<Number> notebookIDRowMapper = new RowMapper<Number>() {
         @Override
-        public Map<String, Object> mapRow(ResultSet rs, int rowNumber) throws SQLException {
-            Map<String, Object> result = new HashMap<String, Object>();
-            result.put(owner_id, rs.getInt(owner_id));
-            result.put(headline, rs.getString(headline));
-            result.put(body_xml, rs.getString(body_xml));
-            result.put(external_id, rs.getString(external_id));
-            result.put(time_stamp, rs.getTimestamp(time_stamp));
-            return result;
+        public Number mapRow(ResultSet rs, int rowNumber) throws SQLException {
+            return rs.getInt(notebook_id);
         }
     };
     
-    
-    /////////////// helpers //////////////////
-    
-    public NewOrExistingSourceInfos contructNewOrExistingSourceInfo(List<SourceInfo> sourceInfoList) {
-        List<NewOrExistingSourceInfo> noeSourceInfoList = new ArrayList<NewOrExistingSourceInfo>();
-        for (SourceInfo sourceInfo : sourceInfoList) {
-            NewOrExistingSourceInfo noeSourceInfo = new NewOrExistingSourceInfo();
-            noeSourceInfo.setSource(sourceInfo);
-            noeSourceInfoList.add(noeSourceInfo);
-        }
-        NewOrExistingSourceInfos result = new NewOrExistingSourceInfos();
-        result.getTarget().addAll(noeSourceInfoList);
-        return result;
-    }
-
 }
